Realized testing language of header only won't work.
Headers may be non-empty or contain statements that are in pure
English even if the rest of the file contains non-English text.
(Imagine a file header containing only a license statement.) Instead
of testing only the header if a header is present, the new code always
tests both header and all comments.